refactor: extract pack step from publish
